\name{fullsample_stratification }
\alias{fullsample_stratification }
\title{fullsample_stratification }
\usage{
fullsample_stratification()
}
\description{
Estimate average treatment effect in each subgroup created from predictionresult.
Whole data is used for making the prediction model.
}
\examples{
fullsample_stratification()
}
